Refactor database schema and remove obsolete migration files
- Updated `env.py` to include new association tables for services and products.
- Deleted obsolete migration files: `929e4d8a7b6f_refactor.py`, `e48cb1b6d34a_refactor_catalog_add_product_and_.py`, and `f17f04027172_refactor_service_category_many_to_many.py`.
- Added new migration file `2df69e15f2a7_init.py` to initialize the database schema with updated tables and relationships.
- Updated `requirements.txt` to include testing dependencies and removed duplicate entries.
- Modified `next.config.ts` to change the Supabase hostname for remote patterns.
- Updated API endpoint in `service.api.ts` to fetch services from the correct path.
