--- back-end/alembic/versions/e48cb1b6d34a_refactor_catalog_add_product_and_.py
@@ -1,275 +1,0 @@
-"""Refactor catalog, add product and treatment plan models
-
-Revision ID: e48cb1b6d34a
-Revises:
-Create Date: 2025-10-03 14:53:59.259777
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel
-
-
-# revision identifiers, used by Alembic.
-revision: str = "e48cb1b6d34a"
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "permission",
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
-        ),
-        sa.Column(
-            "updated_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
-        ),
-        sa.Column("is_deleted", sa.Boolean(), nullable=False),
-        sa.Column("name", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-        sa.Column("description", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_permission_id"), "permission", ["id"], unique=False)
-    op.create_index(
-        op.f("ix_permission_is_deleted"), "permission", ["is_deleted"], unique=False
-    )
-    op.create_index(op.f("ix_permission_name"), "permission", ["name"], unique=True)
-    op.create_table(
-        "role",
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
-        ),
-        sa.Column(
-            "updated_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
-        ),
-        sa.Column("is_deleted", sa.Boolean(), nullable=False),
-        sa.Column("name", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-        sa.Column("description", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_role_id"), "role", ["id"], unique=False)
-    op.create_index(op.f("ix_role_is_deleted"), "role", ["is_deleted"], unique=False)
-    op.create_index(op.f("ix_role_name"), "role", ["name"], unique=True)
-    op.create_table(
-        "service_category",
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
-        ),
-        sa.Column(
-            "updated_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
-        ),
-        sa.Column("is_deleted", sa.Boolean(), nullable=False),
-        sa.Column("name", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-        sa.Column("description", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(
-        op.f("ix_service_category_id"), "service_category", ["id"], unique=False
-    )
-    op.create_index(
-        op.f("ix_service_category_is_deleted"),
-        "service_category",
-        ["is_deleted"],
-        unique=False,
-    )
-    op.create_index(
-        op.f("ix_service_category_name"), "service_category", ["name"], unique=True
-    )
-    op.create_table(
-        "user",
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
-        ),
-        sa.Column(
-            "updated_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
-        ),
-        sa.Column("is_deleted", sa.Boolean(), nullable=False),
-        sa.Column("email", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-        sa.Column("phone", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.Column("full_name", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.Column(
-            "hashed_password", sqlmodel.sql.sqltypes.AutoString(), nullable=False
-        ),
-        sa.Column("is_active", sa.Boolean(), nullable=False),
-        sa.Column("is_superuser", sa.Boolean(), nullable=False),
-        sa.Column("is_email_verified", sa.Boolean(), nullable=False),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_user_email"), "user", ["email"], unique=True)
-    op.create_index(op.f("ix_user_id"), "user", ["id"], unique=False)
-    op.create_index(op.f("ix_user_is_deleted"), "user", ["is_deleted"], unique=False)
-    op.create_index(op.f("ix_user_phone"), "user", ["phone"], unique=True)
-    op.create_table(
-        "default_schedule",
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
-        ),
-        sa.Column(
-            "updated_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
-        ),
-        sa.Column("is_deleted", sa.Boolean(), nullable=False),
-        sa.Column("user_id", sa.Uuid(), nullable=False),
-        sa.Column("day_of_week", sa.Integer(), nullable=False),
-        sa.Column("is_active", sa.Boolean(), nullable=False),
-        sa.Column("start_time", sa.Time(), nullable=True),
-        sa.Column("end_time", sa.Time(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["user_id"],
-            ["user.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(
-        op.f("ix_default_schedule_id"), "default_schedule", ["id"], unique=False
-    )
-    op.create_index(
-        op.f("ix_default_schedule_is_deleted"),
-        "default_schedule",
-        ["is_deleted"],
-        unique=False,
-    )
-    op.create_index(
-        op.f("ix_default_schedule_user_id"),
-        "default_schedule",
-        ["user_id"],
-        unique=False,
-    )
-    op.create_table(
-        "role_permission",
-        sa.Column("role_id", sa.Uuid(), nullable=False),
-        sa.Column("permission_id", sa.Uuid(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["permission_id"],
-            ["permission.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["role_id"],
-            ["role.id"],
-        ),
-        sa.PrimaryKeyConstraint("role_id", "permission_id"),
-    )
-    op.create_table(
-        "service",
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
-        ),
-        sa.Column(
-            "updated_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
-        ),
-        sa.Column("is_deleted", sa.Boolean(), nullable=False),
-        sa.Column("name", sqlmodel.sql.sqltypes.AutoString(length=100), nullable=False),
-        sa.Column("description", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-        sa.Column("price", sa.Float(), nullable=False),
-        sa.Column("duration_minutes", sa.Integer(), nullable=False),
-        sa.Column(
-            "preparation_notes", sqlmodel.sql.sqltypes.AutoString(), nullable=True
-        ),
-        sa.Column(
-            "aftercare_instructions", sqlmodel.sql.sqltypes.AutoString(), nullable=True
-        ),
-        sa.Column(
-            "contraindications", sqlmodel.sql.sqltypes.AutoString(), nullable=True
-        ),
-        sa.Column("category_id", sa.Uuid(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["category_id"],
-            ["service_category.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_service_id"), "service", ["id"], unique=False)
-    op.create_index(
-        op.f("ix_service_is_deleted"), "service", ["is_deleted"], unique=False
-    )
-    op.create_index(op.f("ix_service_name"), "service", ["name"], unique=True)
-    op.create_table(
-        "user_role",
-        sa.Column("user_id", sa.Uuid(), nullable=False),
-        sa.Column("role_id", sa.Uuid(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["role_id"],
-            ["role.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["user_id"],
-            ["user.id"],
-        ),
-        sa.PrimaryKeyConstraint("user_id", "role_id"),
-    )
-    op.create_table(
-        "service_image",
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column(
-            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
-        ),
-        sa.Column(
-            "updated_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
-        ),
-        sa.Column("is_deleted", sa.Boolean(), nullable=False),
-        sa.Column("service_id", sa.Uuid(), nullable=False),
-        sa.Column("url", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-        sa.Column("alt_text", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.Column("is_primary", sa.Boolean(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["service_id"],
-            ["service.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_service_image_id"), "service_image", ["id"], unique=False)
-    op.create_index(
-        op.f("ix_service_image_is_deleted"),
-        "service_image",
-        ["is_deleted"],
-        unique=False,
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_service_image_is_deleted"), table_name="service_image")
-    op.drop_index(op.f("ix_service_image_id"), table_name="service_image")
-    op.drop_table("service_image")
-    op.drop_table("user_role")
-    op.drop_index(op.f("ix_service_name"), table_name="service")
-    op.drop_index(op.f("ix_service_is_deleted"), table_name="service")
-    op.drop_index(op.f("ix_service_id"), table_name="service")
-    op.drop_table("service")
-    op.drop_table("role_permission")
-    op.drop_index(op.f("ix_default_schedule_user_id"), table_name="default_schedule")
-    op.drop_index(op.f("ix_default_schedule_is_deleted"), table_name="default_schedule")
-    op.drop_index(op.f("ix_default_schedule_id"), table_name="default_schedule")
-    op.drop_table("default_schedule")
-    op.drop_index(op.f("ix_user_phone"), table_name="user")
-    op.drop_index(op.f("ix_user_is_deleted"), table_name="user")
-    op.drop_index(op.f("ix_user_id"), table_name="user")
-    op.drop_index(op.f("ix_user_email"), table_name="user")
-    op.drop_table("user")
-    op.drop_index(op.f("ix_service_category_name"), table_name="service_category")
-    op.drop_index(op.f("ix_service_category_is_deleted"), table_name="service_category")
-    op.drop_index(op.f("ix_service_category_id"), table_name="service_category")
-    op.drop_table("service_category")
-    op.drop_index(op.f("ix_role_name"), table_name="role")
-    op.drop_index(op.f("ix_role_is_deleted"), table_name="role")
-    op.drop_index(op.f("ix_role_id"), table_name="role")
-    op.drop_table("role")
-    op.drop_index(op.f("ix_permission_name"), table_name="permission")
-    op.drop_index(op.f("ix_permission_is_deleted"), table_name="permission")
-    op.drop_index(op.f("ix_permission_id"), table_name="permission")
-    op.drop_table("permission")
-    # ### end Alembic commands ###
